refactor(discord): split bot.go into focused files
